pressure sensor working, need to clean up in next commit

+++ Core/Inc/MPL3115A2.h
@@ -8,8 +8,8 @@
 #ifndef INC_MPL3115A2_H_
 #define INC_MPL3115A2_H_
 
+#include <solar_processing_structs.h>
 #include "stm32f4xx_hal.h"
-#include "solar_processing_structs.h"
 
 // MPL3115A2 I2C address
 #define MPL3115A2_ADDR 0x60
@@ -26,7 +26,7 @@
 
 void configure_MPL3115A2(I2C_HandleTypeDef &i2c_device);
 
-double read_baromateric_pressure(I2C_HandleTypeDef &i2c_device);
+float read_baromateric_pressure(I2C_HandleTypeDef &i2c_device);
 
 /* Writes specified register via I2C. Pass in the starting ptr to a uint8_t buffer, with how many bytes you want to write */
 error_code i2c_write_register(I2C_HandleTypeDef &i2c_device, uint32_t reg_addr, uint8_t *data, uint16_t num_bytes);
@@ -35,3 +35,4 @@ error_code i2c_write_register(I2C_HandleTypeDef &i2c_device, uint32_t reg_addr, 
 error_code i2c_read_register(I2C_HandleTypeDef &i2c_device, uint32_t reg_addr, uint8_t *buff, uint16_t num_bytes);
 
 #endif /* INC_MPL3115A2_H_ */
+
